+++ src/am_renderer.cpp
@@ -399,8 +399,8 @@ bool am_render_state::update_state() {
 }
 
 static void setup(am_render_state *rstate, am_framebuffer_id fb,
-    bool clear, glm::vec4 clear_color,
-    int x, int y, int w, int h, int fbw, int fbh, glm::mat4 proj,
+    bool clear, glm::dvec4 clear_color,
+    int x, int y, int w, int h, int fbw, int fbh, glm::dmat4 proj,
     bool has_depthbuffer)
 {
     if (fb != 0) {
@@ -440,8 +440,8 @@ static void setup(am_render_state *rstate, am_framebuffer_id fb,
     rstate->bound_depth_test_state.mask_enabled = has_depthbuffer;
 
     if (clear && (!is_margin ||
-        clear_color.r != 0.0f || clear_color.g != 0.0f ||
-        clear_color.b != 0.0f || clear_color.a != 1.0f))
+        clear_color.r != 0.0 || clear_color.g != 0.0 ||
+        clear_color.b != 0.0 || clear_color.a != 1.0))
     {
         am_set_framebuffer_clear_color(clear_color.r, clear_color.g, clear_color.b, clear_color.a);
         am_clear_framebuffer(true, true, true);
@@ -451,8 +451,8 @@ static void setup(am_render_state *rstate, am_framebuffer_id fb,
 }
 
 void am_render_state::do_render(am_scene_node *root, am_framebuffer_id fb,
-    bool clear, glm::vec4 clear_color, int x, int y, int w, int h, int fbw, int fbh,
-    glm::mat4 proj, bool has_depthbuffer)
+    bool clear, glm::dvec4 clear_color, int x, int y, int w, int h, int fbw, int fbh,
+    glm::dmat4 proj, bool has_depthbuffer)
 {
     setup(this, fb, clear, clear_color, x, y, w, h, fbw, fbh, proj, has_depthbuffer);
     if (root == NULL || am_node_hidden(root)) return;
@@ -798,12 +798,12 @@ static void init_param_name_map(am_render_state *rstate, lua_State *L) {
     lua_pushstring(L, am_conf_default_modelview_matrix_name);
     rstate->modelview_param = &rstate->param_name_map[am_lookup_param_name(L, -1)].value;
     lua_pop(L, 1); // modelview name string
-    rstate->modelview_param->set_mat4(glm::mat4(1.0f));
+    rstate->modelview_param->set_mat4(glm::dmat4(1.0));
 
     lua_pushstring(L, am_conf_default_projection_matrix_name);
     rstate->projection_param = &rstate->param_name_map[am_lookup_param_name(L, -1)].value;
     lua_pop(L, 1); // projection name string
-    rstate->projection_param->set_mat4(glm::mat4(1.0f));
+    rstate->projection_param->set_mat4(glm::dmat4(1.0));
 }
 
 void am_open_renderer_module(lua_State *L) {
